pyproject.toml
[project]
name = "ai-workers"
version = "0.1.0"
description = "StepAI AI Workers - Autonomous agent system"
requires-python = ">=3.11"
dependencies = [
    "slack-bolt>=1.18.0",
    "slack-sdk>=3.27.0",
    "google-api-python-client>=2.100.0",
    "google-auth>=2.23.0",
    "google-auth-oauthlib>=1.2.0",
    "google-auth-httplib2>=0.2.0",
    "anthropic>=0.40.0",
    "apscheduler>=3.10.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "pytz>=2024.1",
    "notion-client>=2.2.0",
    "httpx>=0.27.0",
]

[project.scripts]
hisho = "agents.hisho.main:run"
sanbou = "agents.sanbou.main:run"
sales = "agents.sales.main:run"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
